refactor: implement Updater interface
