+++ data/module/cpu.go
@@ -1,6 +1,7 @@
 package data_module
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
@@ -11,12 +12,63 @@ import (
 	"github.com/timmo001/system-bridge/types"
 )
 
-func (t *Module) UpdateCPUModule() (types.CPUData, error) {
+// CPUFrequency represents CPU frequency information
+type CPUFrequency struct {
+	Current *float64 `json:"current"`
+	Min     *float64 `json:"min"` // TODO: Implement minimum frequency detection
+	Max     *float64 `json:"max"` // TODO: Implement maximum frequency detection
+}
+
+// CPUStats represents CPU statistics
+type CPUStats struct {
+	// TODO: Implement CPU statistics collection
+	CtxSwitches    *int64 `json:"ctx_switches"`    // Context switches count
+	Interrupts     *int64 `json:"interrupts"`      // Hardware interrupts count
+	SoftInterrupts *int64 `json:"soft_interrupts"` // Software interrupts count
+	Syscalls       *int64 `json:"syscalls"`        // System calls count
+}
+
+// CPUTimes represents CPU timing information
+type CPUTimes struct {
+	User      *float64 `json:"user"`
+	System    *float64 `json:"system"`
+	Idle      *float64 `json:"idle"`
+	Interrupt *float64 `json:"interrupt"`
+	DPC       *float64 `json:"dpc"` // TODO: Implement Deferred Procedure Call time tracking
+}
+
+// PerCPU represents per-CPU information
+type PerCPU struct {
+	ID           int           `json:"id"`
+	Frequency    *CPUFrequency `json:"frequency"`
+	Power        *float64      `json:"power"` // TODO: Implement per-CPU power consumption monitoring
+	Times        *CPUTimes     `json:"times"`
+	TimesPercent *CPUTimes     `json:"times_percent"` // TODO: Implement per-CPU time percentage calculations
+	Usage        *float64      `json:"usage"`
+	Voltage      *float64      `json:"voltage"` // TODO: Implement per-CPU voltage monitoring
+}
+
+// CPUData represents overall CPU information
+type CPUData struct {
+	Count        *int          `json:"count"`
+	Frequency    *CPUFrequency `json:"frequency"`
+	LoadAverage  *float64      `json:"load_average"`
+	PerCPU       []PerCPU      `json:"per_cpu"`
+	Power        *float64      `json:"power"` // TODO: Implement overall CPU power consumption monitoring
+	Stats        *CPUStats     `json:"stats"` // TODO: Implement overall CPU statistics collection
+	Temperature  *float64      `json:"temperature"`
+	Times        *CPUTimes     `json:"times"`
+	TimesPercent *CPUTimes     `json:"times_percent"` // TODO: Implement overall CPU time percentage calculations
+	Usage        *float64      `json:"usage"`
+	Voltage      *float64      `json:"voltage"` // TODO: Implement overall CPU voltage monitoring
+}
+
+func (cpuData CPUData) Name() types.ModuleName { return types.ModuleCPU }
+func (cpuData CPUData) Update(ctx context.Context) (any, error) {
 	log.Info("Getting CPU data")
 
-	var cpuData types.CPUData
 	// Initialize arrays
-	cpuData.PerCPU = make([]types.PerCPU, 0)
+	cpuData.PerCPU = make([]PerCPU, 0)
 
 	// Get CPU count
 	count, err := cpu.Counts(true)
@@ -28,7 +80,7 @@ func (t *Module) UpdateCPUModule() (types.CPUData, error) {
 	// Get CPU frequency
 	frequencies, err := cpu.Info()
 	if err == nil && len(frequencies) > 0 {
-		freq := types.CPUFrequency{
+		freq := CPUFrequency{
 			Current: &frequencies[0].Mhz,
 			// TODO: Add implementation for Min frequency
 			// TODO: Add implementation for Max frequency
@@ -38,11 +90,11 @@ func (t *Module) UpdateCPUModule() (types.CPUData, error) {
 
 	// Get per CPU info
 	if len(frequencies) > 0 {
-		perCPU := make([]types.PerCPU, 0, len(frequencies))
+		perCPU := make([]PerCPU, 0, len(frequencies))
 		for i, cpuInfo := range frequencies {
-			perCpuData := types.PerCPU{
+			perCpuData := PerCPU{
 				ID: i,
-				Frequency: &types.CPUFrequency{
+				Frequency: &CPUFrequency{
 					Current: &cpuInfo.Mhz,
 					// TODO: Add implementation for per-CPU Min frequency
 					// TODO: Add implementation for per-CPU Max frequency
@@ -51,7 +103,7 @@ func (t *Module) UpdateCPUModule() (types.CPUData, error) {
 
 			// Get per CPU times
 			if times, err := cpu.Times(true); err == nil && i < len(times) {
-				perCpuData.Times = &types.CPUTimes{
+				perCpuData.Times = &CPUTimes{
 					User:      &times[i].User,
 					System:    &times[i].System,
 					Idle:      &times[i].Idle,
@@ -78,7 +130,7 @@ func (t *Module) UpdateCPUModule() (types.CPUData, error) {
 
 	// Get overall CPU times
 	if times, err := cpu.Times(false); err == nil && len(times) > 0 {
-		cpuData.Times = &types.CPUTimes{
+		cpuData.Times = &CPUTimes{
 			User:      &times[0].User,
 			System:    &times[0].System,
 			Idle:      &times[0].Idle,
