+++ data/module/displays.go
@@ -1,16 +1,20 @@
 package data_module
 
 import (
+	"context"
+
 	"github.com/charmbracelet/log"
 	"github.com/timmo001/system-bridge/data/module/displays"
 	"github.com/timmo001/system-bridge/types"
 )
 
-func (t *Module) UpdateDisplaysModule() (types.DisplaysData, error) {
-	log.Info("Getting displays data")
+// DisplaysData represents information about all display devices
+type DisplaysData struct {
+}
 
-	var displaysData types.DisplaysData
-	displaysData = make([]types.Display, 0)
+func (displaysData DisplaysData) Name() types.ModuleName { return types.ModuleDisplays }
+func (displaysData DisplaysData) Update(ctx context.Context) (any, error) {
+	log.Info("Getting displays data")
 
 	displays, err := displays.GetDisplays()
 	if err != nil {
@@ -18,6 +22,5 @@ func (t *Module) UpdateDisplaysModule() (types.DisplaysData, error) {
 		return displaysData, err
 	}
 
-	displaysData = displays
-	return displaysData, nil
+	return displays, nil
 }
