+++ data/module/displays/displays_windows.go
@@ -10,12 +10,11 @@ import (
 	"unsafe"
 
 	"github.com/charmbracelet/log"
-	"github.com/timmo001/system-bridge/types"
 )
 
 var (
-	user32               = syscall.NewLazyDLL("user32.dll")
-	enumDisplayMonitors  = user32.NewProc("EnumDisplayMonitors")
+	user32              = syscall.NewLazyDLL("user32.dll")
+	enumDisplayMonitors = user32.NewProc("EnumDisplayMonitors")
 	getMonitorInfo      = user32.NewProc("GetMonitorInfoW")
 	displayPathRegex    = regexp.MustCompile(`^\\\\.\\|^\\\\\?\\`)
 	displayNameRegex    = regexp.MustCompile(`^DISPLAY(\d+)$`)
@@ -26,10 +25,10 @@ type RECT struct {
 }
 
 type MONITORINFOEX struct {
-	CbSize    uint32
-	Monitor   RECT
-	WorkArea  RECT
-	Flags     uint32
+	CbSize     uint32
+	Monitor    RECT
+	WorkArea   RECT
+	Flags      uint32
 	DeviceName [32]uint16
 }
 
@@ -41,8 +40,8 @@ func formatDisplayName(id string) string {
 	return id
 }
 
-func getDisplays() ([]types.Display, error) {
-	var displays []types.Display
+func getDisplays() ([]Display, error) {
+	var displays []Display
 
 	callback := syscall.NewCallback(func(handle syscall.Handle, dc syscall.Handle, rect *RECT, data uintptr) uintptr {
 		var info MONITORINFOEX
@@ -63,13 +62,13 @@ func getDisplays() ([]types.Display, error) {
 		pixelClock := 0.0
 		refreshRate := 0.0
 
-    displayName := syscall.UTF16ToString(info.DeviceName[:])
+		displayName := syscall.UTF16ToString(info.DeviceName[:])
 		cleanID := displayPathRegex.ReplaceAllString(displayName, "")
-		display := types.Display{
+		display := Display{
 			ID:                   cleanID,
 			Name:                 formatDisplayName(cleanID),
 			ResolutionHorizontal: int(width),
-			ResolutionVertical:  int(height),
+			ResolutionVertical:   int(height),
 			X:                    int(info.Monitor.Left),
 			Y:                    int(info.Monitor.Top),
 			IsPrimary:            &isPrimary,
