+++ data/module/media/media_windows.go
@@ -3,11 +3,7 @@
 
 package media
 
-import (
-	"github.com/timmo001/system-bridge/types"
-)
-
-func getMediaData(mediaData types.MediaData) (types.MediaData, error) {
-  // TODO: Implement
-  return mediaData, nil
+func getMediaData(media Media) (Media, error) {
+	// TODO: Implement
+	return media, nil
 }
