+++ data/module/sensors/sensors_other.go
@@ -3,16 +3,12 @@
 
 package sensors
 
-import (
-	"github.com/timmo001/system-bridge/types"
-)
-
-func getWindowsSensorsData() (*types.SensorsWindows, error) {
-	var sensorsData types.SensorsWindows
-	sensorsData.Hardware = make([]types.SensorsWindowsHardware, 0)
-	sensorsData.NVIDIA = &types.SensorsNVIDIA{
-		Displays: make([]types.SensorsNVIDIADisplay, 0),
-		GPUs:     make([]types.SensorsNVIDIAGPU, 0),
+func GetWindowsSensorsData() (*SensorsWindows, error) {
+	var sensorsData SensorsWindows
+	sensorsData.Hardware = make([]SensorsWindowsHardware, 0)
+	sensorsData.NVIDIA = &SensorsNVIDIA{
+		Displays: make([]SensorsNVIDIADisplay, 0),
+		GPUs:     make([]SensorsNVIDIAGPU, 0),
 	}
 
 	return &sensorsData, nil
