--- types/gpus.go
@@ -1,20 +1,0 @@
-package types
-
-// GPU represents information about a GPU device
-type GPU struct {
-	ID          string   `json:"id"`
-	Name        string   `json:"name"`
-	CoreClock   *float64 `json:"core_clock"`
-	CoreLoad    *float64 `json:"core_load"`
-	FanSpeed    *float64 `json:"fan_speed"`
-	MemoryClock *float64 `json:"memory_clock"`
-	MemoryLoad  *float64 `json:"memory_load"`
-	MemoryFree  *float64 `json:"memory_free"`
-	MemoryUsed  *float64 `json:"memory_used"`
-	MemoryTotal *float64 `json:"memory_total"`
-	PowerUsage  *float64 `json:"power_usage"`
-	Temperature *float64 `json:"temperature"`
-}
-
-// GPUsData represents information about all GPU devices
-type GPUsData = []GPU
