+++ types/module.go
@@ -8,7 +8,7 @@ const (
 	ModuleCPU       ModuleName = "cpu"
 	ModuleDisks     ModuleName = "disks"
 	ModuleDisplays  ModuleName = "displays"
-	ModuleGPUs      ModuleName = "gpus"
+	ModuleGPU       ModuleName = "gpus"
 	ModuleMedia     ModuleName = "media"
 	ModuleMemory    ModuleName = "memory"
 	ModuleNetworks  ModuleName = "networks"
@@ -16,10 +16,3 @@ const (
 	ModuleSensors   ModuleName = "sensors"
 	ModuleSystem    ModuleName = "system"
 )
-
-// Module represents a data module
-type Module struct {
-	Module  ModuleName `json:"module" mapstructure:"module"`
-	Data    any        `json:"data" mapstructure:"data"`
-	Updated string     `json:"updated" mapstructure:"updated"`
-}
